added jwt

+++ resources/books/routes.py
@@ -2,17 +2,17 @@ from flask import request
 from uuid import uuid4
 from flask.views import MethodView
 from flask_smorest import abort
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
-from db import books
 from . import bp
-from schemas import BookSchema
-from models.BookModel import BookModel
+from schemas import BookSchema, BookSchemaNested
+from models.UserModel import BookModel
 
 
 @bp.route('/<book_id>')
 class Book(MethodView):
 
-    @bp.response(200, BookSchema)
+    @bp.response(200, BookSchemaNested)
     def get(self, book_id):
         book = BookModel.query.get(book_id)
         if book:
@@ -20,11 +20,11 @@ class Book(MethodView):
         else:
             abort(400, message='Book not found')
  
-
+    @jwt_required()
     @bp.arguments(BookSchema)
     def put(self, book_data, book_id):
         book = BookModel.query.get(book_id)
-        if book:
+        if book and book.user_id == get_jwt_identity():
             book.title = book_data['title']
             book.author = book_data['author']
             book.publisher = book_data['publisher']
@@ -34,11 +34,10 @@ class Book(MethodView):
         return {'Message': 'Invalid Book Id'}, 400
         
      
-   
-
+    @jwt_required()
     def delete(self, book_id):
         book = BookModel.query.get(book_id)
-        if book:
+        if book and book.user_id == get_jwt_identity():
             book.delete()
             return {'Message': f"Book: {book.title} deleted"}, 202
         return {'Message': 'Invalid book'}, 400
@@ -51,10 +50,12 @@ class BookList(MethodView):
     def get(self):
         return BookModel.query.all()
     
+    @jwt_required()
     @bp.arguments(BookSchema)
     def post(self, book_data):
         try:
             book = BookModel()
+            book.user_id = get_jwt_identity()
             book.title = book_data['title']
             book.author = book_data['author']
             book.publisher = book_data['publisher']
